Increased single precision Delay fluctuations
Single precision being between one frame to the next. An overflow beyond the delay of one frame but not so great that it exceedes two frames would yield an underflow of time such that a framerate would run faster than the specified time.

Single frame is the maximum precision due to other implementations being in the scope of delta which is outside of the scope of Delay

Added static offset conversion of Timestamp values to raw llu